Port to PyGObject
Convert function and method call to new API.

Rename first parameter of size and draw functions to reflect its
actual type. The first parameter of both size and draw functions
is in fact a Cairo context, not an image.

Skip creating a context in both size_of_Token() and draw_Token().
We have one already.